Fixed loading bug, made harvester somewhat generic.
