+++ src/main/java/rs/teslaris/core/controller/OAIPMHHarvestController.java
@@ -4,17 +4,17 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import rs.teslaris.core.harvester.OAIPMHHarvesterTest;
+import rs.teslaris.core.harvester.OAIPMHHarvester;
 
 @RestController
 @RequestMapping("/api/harvest")
 @RequiredArgsConstructor
 public class OAIPMHHarvestController {
 
-    private final OAIPMHHarvesterTest oaipmhHarvesterTest;
+    private final OAIPMHHarvester oaipmhHarvester;
 
     @GetMapping
     public void testHarvestForOU() {
-        oaipmhHarvesterTest.harvest();
+        oaipmhHarvester.harvest("openaire_cris_orgunits");
     }
 }
