+++ src/main/java/rs/teslaris/core/harvester/organisationunits/Record.java
@@ -1,4 +1,4 @@
-package rs.teslaris.core.harvester;
+package rs.teslaris.core.harvester.organisationunits;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -7,6 +7,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 import lombok.Getter;
 import lombok.Setter;
+import rs.teslaris.core.harvester.common.Header;
 
 @XmlType(name = "TRecord")
 @XmlAccessorType(XmlAccessType.FIELD)
